tmp: Removed unnecessary files when running tests with phpunit command.
