--- lib/Cake/Console/Templates/skel/webroot/test.php
@@ -1,96 +1,0 @@
-<?php
-/**
- * Web Access Frontend for TestSuite
- *
- * @link          https://book.cakephp.org/2.0/en/development/testing.html
- * @package       app.webroot
- * @since         CakePHP(tm) v 1.2.0.4433
- */
-
-set_time_limit(0);
-ini_set('display_errors', 1);
-
-/**
- * Use the DS to separate the directories in other defines
- */
-if (!defined('DS')) {
-	define('DS', DIRECTORY_SEPARATOR);
-}
-
-/**
- * These defines should only be edited if you have CakePHP installed in
- * a directory layout other than the way it is distributed.
- * When using custom settings be sure to use the DS and do not add a trailing DS.
- */
-
-/**
- * The full path to the directory which holds "app", WITHOUT a trailing DS.
- */
-if (!defined('ROOT')) {
-	define('ROOT', dirname(dirname(dirname(__FILE__))));
-}
-
-/**
- * The actual directory name for the "app".
- */
-if (!defined('APP_DIR')) {
-	define('APP_DIR', basename(dirname(dirname(__FILE__))));
-}
-
-/**
- * The absolute path to the "Cake" directory, WITHOUT a trailing DS.
- *
- * For ease of development CakePHP uses PHP's include_path. If you
- * need to cannot modify your include_path, you can set this path.
- *
- * Leaving this constant undefined will result in it being defined in Cake/bootstrap.php
- *
- * The following line differs from its sibling
- * /app/webroot/test.php
- */
-//define('CAKE_CORE_INCLUDE_PATH', __CAKE_PATH__);
-
-/**
- * This auto-detects CakePHP as a composer installed library.
- * You may remove this if you are not planning to use composer (not recommended, though).
- */
-$vendorPath = ROOT . DS . APP_DIR . DS . 'Vendor' . DS . 'cakephp' . DS . 'cakephp' . DS . 'lib';
-$dispatcher = 'Cake' . DS . 'Console' . DS . 'ShellDispatcher.php';
-if (!defined('CAKE_CORE_INCLUDE_PATH') && file_exists($vendorPath . DS . $dispatcher)) {
-	define('CAKE_CORE_INCLUDE_PATH', $vendorPath);
-}
-
-/**
- * Editing below this line should not be necessary.
- * Change at your own risk.
- */
-if (!defined('WEBROOT_DIR')) {
-	define('WEBROOT_DIR', basename(dirname(__FILE__)));
-}
-if (!defined('WWW_ROOT')) {
-	define('WWW_ROOT', dirname(__FILE__) . DS);
-}
-
-if (!defined('CAKE_CORE_INCLUDE_PATH')) {
-	if (function_exists('ini_set')) {
-		ini_set('include_path', ROOT . DS . 'lib' . PATH_SEPARATOR . ini_get('include_path'));
-	}
-	if (!include 'Cake' . DS . 'bootstrap.php') {
-		$failed = true;
-	}
-} else {
-	if (!include CAKE_CORE_INCLUDE_PATH . DS . 'Cake' . DS . 'bootstrap.php') {
-		$failed = true;
-	}
-}
-if (!empty($failed)) {
-	trigger_error("CakePHP core could not be found. Check the value of CAKE_CORE_INCLUDE_PATH in APP/webroot/index.php. It should point to the directory containing your " . DS . "cake core directory and your " . DS . "vendors root directory.", E_USER_ERROR);
-}
-
-if (Configure::read('debug') < 1) {
-	throw new NotFoundException(__d('cake_dev', 'Debug setting does not allow access to this URL.'));
-}
-
-require_once CAKE . 'TestSuite' . DS . 'CakeTestSuiteDispatcher.php';
-
-CakeTestSuiteDispatcher::run();
