+++ lib/Cake/Test/bootstrap.php
@@ -7,15 +7,7 @@ if (!defined('DS')) {
 	define('DS', DIRECTORY_SEPARATOR);
 }
 require_once __DIR__ . DS . 'bootstrap' . DS . 'cake_dot_php.php';
-
-/*
- * loading of lib/Cake/TestSuite/CakeTestSuiteDispatcher.php
- * In bootstrap.php, it is sufficient if the const(s) are defined outside the class of CakeTestSuiteDispatcher.php.
- * However, when loading CakeTestSuiteDispatcher.php in the unit test, a double definition of const(s) error occurs,
- * so load it here.
- */
-App::uses('CakeTestSuiteDispatcher', 'TestSuite');
-App::load('CakeTestSuiteDispatcher');
+require_once __DIR__ . DS . 'bootstrap' . DS . 'cake_test_suite_const.php';
 
 /*
  * Classes that can be used without declaring App::uses()
